requirements.txt
phidata
openai
lancedb 
tantivy 
pypdf 
sqlalchemy 
pgvector 
psycopg[binary]